// frontend/LawyerSite.Web/src/app/services/client/web-api-client.ts
export interface AppointmentResponseDto {
    id?: number;
    clientName?: string;
    clientEmail?: string;
    clientPhone?: string;
    appointmentTime?: Date;
    durationMinutes?: number;
    notes?: string | undefined;
    status?: string;
    createdAt?: Date;
    confirmedAt?: Date | undefined;
    cancelledAt?: Date | undefined;
    isBlocked?: boolean;
    blockReason?: string | undefined;
    serviceId?: number | undefined;
    serviceTitle?: string | undefined;
}
// ...
export interface CreateAppointmentDto {
    clientName?: string;
    clientEmail?: string;
    clientPhone?: string;
    appointmentTime?: Date;
    notes?: string | undefined;
    serviceId?: number | undefined;
}